+++ client/src/Commands/uploadToServer.ts
@@ -46,16 +46,16 @@ export async function run(args: any) {
   // Override the "deleteRemote" setting for safety.
   config["deleteRemote"] = false;
 
-  // Make sure the path to upload is relative to avoid uploading the whole disk.
-  let workspaceFolders = Workspace.workspaceFolders;
-  if (workspaceFolders === undefined) {
-    window.showWarningMessage(
-      "No workspace or folder found. \n Please open the folder containing your .sp file, not just the .sp file."
-    );
-    return 1;
+  // Concat the workspace with it's root if the path is relative.
+  let workspaceRoot = workspaceFolder.uri.fsPath;
+  if (config["isRootRelative"]) {
+    config["localRoot"] = join(workspaceRoot, config["localRoot"]);
   }
-  let workspaceRoot: string = workspaceFolders[0].uri.fsPath;
-  config["localRoot"] = join(workspaceRoot, config["localRoot"]);
+
+  // Delete that setting to avoid problems with the ftp/sftp library
+  delete config["isRootRelative"];
+
+  console.log("Starting the upload");
   ftpDeploy
     .deploy(config)
     .then(() => {